Use app's own storage dir to solve permissions
SDK 29 is the last version that supports access to arbitrary files on the device. THis means we can no longer use a folder in the "Pictures" directory. This commit makes the app use its own storage directory (/storage/emulated/0/Android/data/info.androidhive.androidcamera/files/) instead.

I clearly still need to fix the actual name of the app, but this is progress.