Implement full i18n
All text is now internationalised. In addition, we no longer use an ugly function that parses the json file for every single string to translate. Instead, we load the i18n json once and parse that into a Hashmap.

We should do something similar for the other jsons (game settings, etc.) but this will have to wait for another time